requirements.txt
alpaca-trade-api==0.53.0
autopep8==2.0.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.1.3
Flask==2.2.3
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
multitasking==0.0.11
newsapi-python==0.2.7
nltk==3.5
numpy==1.24.2
pandas==1.5.3
pycodestyle==2.10.0
python-dateutil==2.8.1
pytz==2021.1
regex==2022.10.31
requests==2.25.1
six==1.15.0
soupsieve==2.2
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.5
websocket-client==0.57.0
websockets==8.1
Werkzeug==2.2.3
yfinance==0.1.55
